--- src/main/scala/scalaBot/EventStorageRepo.scala
@@ -1,13 +1,0 @@
-package scalaBot
-
-trait EventStorageRepo {
-  def addEvent(t: Event): Unit
-
-  def deleteEvent(t: Event): Unit
-
-  def changeEvent(t: Event): Unit
-
-  def findEventById(ind: Int): Event
-
-  def toList: List[Event]
-}
