--- src/main/scala/scalaBot/Registration.scala
@@ -1,11 +1,0 @@
-package scalaBot
-
-trait Registration {
-  // working with memory
-  /* register someone, return success of operation */
-  def registration(): Boolean
-
-  // working with memory
-  /* delete user, return success of operation */
-  def delete(): Boolean
-}
