+++ src/main/scala/scalaBot/repository/EventStorageRepository.scala
@@ -1,0 +1,7 @@
+package scalaBot.repository
+
+import scalaBot.Event
+
+class EventStorageRepository extends HashMapRepository[Event] {
+
+}
