+++ src/main/scala/scalaBot/repository/EventUsersRepository.scala
@@ -1,0 +1,16 @@
+package scalaBot.repository
+
+import scalaBot.User
+
+class EventUsersRepository extends HashMapRepository[List[User]] {
+
+  def addOne(id: Int, t: User): Unit = table.updateWith(id) {
+    case Some(list) => Some(list :+ t)
+    case None       => Some(List(t))
+  }
+
+  def deleteOne(id: Int, user: User): Unit = table.updateWith(id) {
+    case Some(list) => Some(list.filterNot(_ == user))
+    case None       => None
+  }
+}
