+++ src/main/scala/scalaBot/repository/UserRepository.scala
@@ -1,0 +1,6 @@
+package scalaBot.repository
+
+import scalaBot.User
+
+class UserRepository extends HashMapRepository[User]  {
+}
